Move Potato and PotatoField from clustering to artifactdetection
