+++ examples/simulated/plot_mean_median_comparison.py
@@ -15,7 +15,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from sklearn.datasets import make_blobs
 
-from pyriemann.clustering import Potato
+from pyriemann.artifactdetection import Potato
 from pyriemann.datasets import make_outliers
 from pyriemann.utils import (
     mean_euclid,
